Add the ability to turn subcommands off for a project
A new element in the manifest XML, disabled-subcmds, can turn off subcmds
project-wide that don't make sense or should not be run (e.g. turning off
"upload" for projects that do not use Gerrit). For example:

    <disabled-subcmds subcmds="upload" />

Will prevent the "upload" subcommand from being run or showing up in the help
and usage.